Extract DataInput to a separate file
